+++ Scenes/Main/main.gd
@@ -8,7 +8,7 @@ var socket: WebSocketPeer = WebSocketPeer.new()
 const MAX_MESSAGE_LENGTH: int = 2000
 const HE6_HALF_LIFE: float = 0.8067
 
-@onready var callsign_node: CallsignSelector = $MarginContainer/MainContainer/Body/Sidebar/CallsignContainer/CallsignEdit
+@onready var callsign_node: CallsignSelector = $Body/Sidebar/CallsignContainer/CallsignEdit
 var callsign: int = 0:
 	set(value):
 		if value < 0:
@@ -44,6 +44,9 @@ static func on_nicknames_reload() -> void:
 signal reload_image_inversion
 static func on_image_inversion_reload() -> void:
 	instance.reload_image_inversion.emit()
+signal reload_library
+static func on_library_reload() -> void:
+	instance.reload_library.emit()
 
 signal connected_user_change
 
@@ -215,7 +218,7 @@ func send_message(written: String) -> Array:
 	var strig: Array = sig.map(func (a): return str(a)) as Array[String]
 	# extra check here to account for auto-channel-prefix
 	if strig.size() > MAX_MESSAGE_LENGTH:
-		Chat.new_log(Chat.State.INPUT_TOO_LONG)
+		Chat.new_log(Chat.State.INPUT_TOO_LONG, [strig.size()])
 		return [false, MessageCompilationResult.MESSAGE_FAILED]
 	strig.insert(0, "M")
 	var o: String = ",".join(strig)
@@ -327,7 +330,6 @@ func _on_dictionary_save_open_pressed():
 
 func _on_dsve_button_pressed():
 	OS.shell_open("https://dsve.akqqa.dev/")
-	SoundManager.play_sound(SoundManager.Sounds.CLICK)
 
 func _on_callsign_edit_callsign_submitted(new_value: int) -> void:
 	callsign = new_value