+++ Scenes/Main/sidebar_tabber.gd
@@ -5,6 +5,6 @@ func _ready():
 	current_tab = 0
 
 func refresh():
-	self.set("tab_0/title", DictionaryHandler.get_or_default_signal_name(-42))
-	self.set("tab_1/title", DictionaryHandler.get_or_default_signal_name(-130))
-	self.set("tab_2/title", DictionaryHandler.get_or_default_signal_name(-241))
+	set_tab_title(0, DictionaryHandler.get_or_default_signal_name(-42))
+	set_tab_title(1, DictionaryHandler.get_or_default_signal_name(-130))
+	set_tab_title(2, DictionaryHandler.get_or_default_signal_name(-241))